fix control jacobian size issue
